Extract DelegatingSchedulerSharedLister to separate file
